loading_widget  and fix import
