+++ lib/presentation/random_details/random_details_screen.dart
@@ -5,8 +5,8 @@ import 'package:NeQuo/domain/usecases/share_quote.dart';
 import 'package:NeQuo/presentation/random_details/bloc/random_details_bloc.dart';
 import 'package:NeQuo/presentation/random_details/bloc/random_details_event.dart';
 import 'package:NeQuo/presentation/random_details/bloc/random_details_state.dart';
-import 'package:NeQuo/presentation/shared/action_button.dart';
-import 'package:NeQuo/presentation/shared/favorite_bloc.dart';
+import 'package:NeQuo/presentation/shared/widgets/action_button.dart';
+import 'package:NeQuo/presentation/shared/bloc/favorite_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
