+++ lib/presentation/shared/bloc/favorite_bloc.dart
