+++ lib/presentation/shared/widgets/loading_widget.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class LoadingWidget extends StatelessWidget {
+  final Key key;
+
+  LoadingWidget(this.key) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: CircularProgressIndicator(),
+    );
+  }
+}
